requirements.txt
# Web Scraper Agent — Python Dependencies
# Install with: pip install -r requirements.txt
# Then run: playwright install chromium

groq>=0.9.0
playwright>=1.44.0
trafilatura>=1.9.0
pydantic>=2.7.0
pandas>=2.2.0
openpyxl>=3.1.0          # Excel export
rich>=13.7.0              # Beautiful CLI tables
streamlit>=1.35.0         # Web UI
beautifulsoup4>=4.12.0    # HTML parsing fallback
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0